Change os image to `latest` for precompiling NIF
This commit want to change OS image in nif_precompile on GitHub Actions.
As we experienced in #104, a boring issue can arise when the version provided by GHA becomes deprecated. Since this repository is not so sensitive to OS versions for compiling NIFs, it is acceptable to fix it to `-latest`.